Cleanup SQLAlchemy column creation, fix static SRID assignment
Since all SQLAlchemy factory logic is found in `schematools.factories`,
this should also include the column type data.
The columns all received the same SRID, and never looked at what the
actual `DatasetSchemaField.srid` type was!
